+++ exporters/trace/jaeger/internal/gen-go/zipkincore/zipkin_collector-remote/zipkin_collector-remote.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"go.opentelemetry.io/otel/exporters/trace/jaeger/internal/gen-go/jaeger"
+	"go.opentelemetry.io/otel/exporters/trace/jaeger/internal/gen-go/zipkincore"
 	"go.opentelemetry.io/otel/exporters/trace/jaeger/internal/third_party/thrift/lib/go/thrift"
 	"math"
 	"net"
@@ -16,13 +16,13 @@ import (
 	"strings"
 )
 
-var _ = jaeger.GoUnusedProtection__
+var _ = zipkincore.GoUnusedProtection__
 
 func Usage() {
 	fmt.Fprintln(os.Stderr, "Usage of ", os.Args[0], " [-h host:port] [-u url] [-f[ramed]] function [arg1 [arg2...]]:")
 	flag.PrintDefaults()
 	fmt.Fprintln(os.Stderr, "\nFunctions:")
-	fmt.Fprintln(os.Stderr, "   submitBatches( batches)")
+	fmt.Fprintln(os.Stderr, "   submitZipkinBatch( spans)")
 	fmt.Fprintln(os.Stderr)
 	os.Exit(0)
 }
@@ -137,37 +137,37 @@ func main() {
 	}
 	iprot := protocolFactory.GetProtocol(trans)
 	oprot := protocolFactory.GetProtocol(trans)
-	client := jaeger.NewCollectorClient(thrift.NewTStandardClient(iprot, oprot))
+	client := zipkincore.NewZipkinCollectorClient(thrift.NewTStandardClient(iprot, oprot))
 	if err := trans.Open(); err != nil {
 		fmt.Fprintln(os.Stderr, "Error opening socket to ", host, ":", port, " ", err)
 		os.Exit(1)
 	}
 
 	switch cmd {
-	case "submitBatches":
+	case "submitZipkinBatch":
 		if flag.NArg()-1 != 1 {
-			fmt.Fprintln(os.Stderr, "SubmitBatches requires 1 args")
+			fmt.Fprintln(os.Stderr, "SubmitZipkinBatch requires 1 args")
 			flag.Usage()
 		}
-		arg19 := flag.Arg(1)
-		mbTrans20 := thrift.NewTMemoryBufferLen(len(arg19))
-		defer mbTrans20.Close()
-		_, err21 := mbTrans20.WriteString(arg19)
-		if err21 != nil {
+		arg11 := flag.Arg(1)
+		mbTrans12 := thrift.NewTMemoryBufferLen(len(arg11))
+		defer mbTrans12.Close()
+		_, err13 := mbTrans12.WriteString(arg11)
+		if err13 != nil {
 			Usage()
 			return
 		}
-		factory22 := thrift.NewTJSONProtocolFactory()
-		jsProt23 := factory22.GetProtocol(mbTrans20)
-		containerStruct0 := jaeger.NewCollectorSubmitBatchesArgs()
-		err24 := containerStruct0.ReadField1(context.Background(), jsProt23)
-		if err24 != nil {
+		factory14 := thrift.NewTJSONProtocolFactory()
+		jsProt15 := factory14.GetProtocol(mbTrans12)
+		containerStruct0 := zipkincore.NewZipkinCollectorSubmitZipkinBatchArgs()
+		err16 := containerStruct0.ReadField1(context.Background(), jsProt15)
+		if err16 != nil {
 			Usage()
 			return
 		}
-		argvalue0 := containerStruct0.Batches
+		argvalue0 := containerStruct0.Spans
 		value0 := argvalue0
-		fmt.Print(client.SubmitBatches(context.Background(), value0))
+		fmt.Print(client.SubmitZipkinBatch(context.Background(), value0))
 		fmt.Print("\n")
 		break
 	case "":